requirements.txt
# --- Core Framework ---
Django==4.2.16
django-cors-headers==3.7.0
django-controlcenter==0.3.2

# --- Database (Oracle) ---
cx_Oracle==8.3.0
oracledb==2.2.0

# --- Report & Excel ---
# Dùng trong admin_volt/models.py và utils.py để xuất báo cáo [cite: 68]
openpyxl==3.1.5

# --- Kafka Integration ---
# Dùng trong store/models.py (SalesProducer) [cite: 119]
kafka-python==2.2.6

# --- Utilities ---
# Dùng cho models.ImageField trong store/models.py [cite: 104]
pillow==10.3.0
# Quản lý biến môi trường (Good practice)
python-dotenv==1.1.1
# Dùng để chạy server ASGI (do có file asgi.py) [cite: 217]
uvicorn==0.38.0
gunicorn==21.2.0
whitenoise==6.6.0

six==1.16.0


# --- Testing ---
pytest==8.3.4